+++ anyblok/plugins.py
@@ -1,6 +1,9 @@
 from nose.plugins import Plugin
 from anyblok.config import Configuration
-from anyblok import load_init_function_from_entry_points
+from anyblok import (
+    load_init_function_from_entry_points,
+    configuration_post_load,
+)
 from anyblok.blok import BlokManager
 from anyblok.registry import RegistryManager
 from os.path import join, exists
@@ -70,10 +73,11 @@ class AnyBlokPlugin(Plugin):
             # Load the registry here not in configuration,
             # because the configuration are not load in order of score
             self.registryLoaded = True
-            BlokManager.load()
             load_init_function_from_entry_points(unittest=True)
             Configuration.load_config_for_test()
             Configuration.parse_options(self.AnyBlokOptions, ('bloks',))
+            configuration_post_load()
+            BlokManager.load()
             db_name = Configuration.get('db_name')
             if not db_name:
                 raise Exception("No database defined to run Test")